Fixed deprecated references to HSSFSheet.getRow(short) etc.  Removed all deprecated warnings from hssf.usermodel.examples.*
git-svn-id: https://svn.apache.org/repos/asf/poi/trunk@740179 13f79535-47bb-0310-9956-ffa450edef68

+++ src/examples/src/org/apache/poi/hssf/usermodel/examples/Alignment.java
@@ -29,27 +29,23 @@ import java.io.IOException;
  *
  * @author Glen Stampoultzis (glens at apache.org)
  */
-public class Alignment
-{
-    public static void main(String[] args)
-            throws IOException
-    {
+public class Alignment {
+    public static void main(String[] args) throws IOException {
         HSSFWorkbook wb = new HSSFWorkbook();
         HSSFSheet sheet = wb.createSheet("new sheet");
-        HSSFRow row = sheet.createRow((short) 2);
-        createCell(wb, row, (short) 0, HSSFCellStyle.ALIGN_CENTER);
-        createCell(wb, row, (short) 1, HSSFCellStyle.ALIGN_CENTER_SELECTION);
-        createCell(wb, row, (short) 2, HSSFCellStyle.ALIGN_FILL);
-        createCell(wb, row, (short) 3, HSSFCellStyle.ALIGN_GENERAL);
-        createCell(wb, row, (short) 4, HSSFCellStyle.ALIGN_JUSTIFY);
-        createCell(wb, row, (short) 5, HSSFCellStyle.ALIGN_LEFT);
-        createCell(wb, row, (short) 6, HSSFCellStyle.ALIGN_RIGHT);
+        HSSFRow row = sheet.createRow(2);
+        createCell(wb, row, 0, HSSFCellStyle.ALIGN_CENTER);
+        createCell(wb, row, 1, HSSFCellStyle.ALIGN_CENTER_SELECTION);
+        createCell(wb, row, 2, HSSFCellStyle.ALIGN_FILL);
+        createCell(wb, row, 3, HSSFCellStyle.ALIGN_GENERAL);
+        createCell(wb, row, 4, HSSFCellStyle.ALIGN_JUSTIFY);
+        createCell(wb, row, 5, HSSFCellStyle.ALIGN_LEFT);
+        createCell(wb, row, 6, HSSFCellStyle.ALIGN_RIGHT);
 
         // Write the output to a file
         FileOutputStream fileOut = new FileOutputStream("workbook.xls");
         wb.write(fileOut);
         fileOut.close();
-
     }
 
     /**
@@ -60,12 +56,11 @@ public class Alignment
      * @param column    the column number to create the cell in
      * @param align     the alignment for the cell.
      */
-    private static void createCell(HSSFWorkbook wb, HSSFRow row, short column, short align)
-    {
+    private static void createCell(HSSFWorkbook wb, HSSFRow row, int column, int align) {
         HSSFCell cell = row.createCell(column);
         cell.setCellValue("Align It");
         HSSFCellStyle cellStyle = wb.createCellStyle();
-        cellStyle.setAlignment(align);
+        cellStyle.setAlignment((short)align);
         cell.setCellStyle(cellStyle);
     }
 }
