+++ src/examples/src/org/apache/poi/hssf/usermodel/examples/CreateCells.java
@@ -1,4 +1,3 @@
-
 /* ====================================================================
    Licensed to the Apache Software Foundation (ASF) under one or more
    contributor license agreements.  See the NOTICE file distributed with
@@ -15,7 +14,6 @@
    See the License for the specific language governing permissions and
    limitations under the License.
 ==================================================================== */
-        
 
 package org.apache.poi.hssf.usermodel.examples;
 
@@ -32,24 +30,21 @@ import java.io.IOException;
  *
  * @author Glen Stampoultzis (glens at apache.org)
  */
-public class CreateCells
-{
-    public static void main(String[] args)
-        throws IOException
-    {
+public class CreateCells {
+    public static void main(String[] args) throws IOException {
         HSSFWorkbook wb = new HSSFWorkbook();
         HSSFSheet sheet = wb.createSheet("new sheet");
 
         // Create a row and put some cells in it. Rows are 0 based.
-        HSSFRow row = sheet.createRow((short)0);
+        HSSFRow row = sheet.createRow(0);
         // Create a cell and put a value in it.
-        HSSFCell cell = row.createCell((short)0);
+        HSSFCell cell = row.createCell(0);
         cell.setCellValue(1);
 
         // Or do it on one line.
-        row.createCell((short)1).setCellValue(1.2);
-        row.createCell((short)2).setCellValue("This is a string");
-        row.createCell((short)3).setCellValue(true);
+        row.createCell(1).setCellValue(1.2);
+        row.createCell(2).setCellValue("This is a string");
+        row.createCell(3).setCellValue(true);
 
         // Write the output to a file
         FileOutputStream fileOut = new FileOutputStream("workbook.xls");
