+++ src/examples/src/org/apache/poi/hssf/usermodel/examples/CreateDateCells.java
@@ -1,4 +1,3 @@
-
 /* ====================================================================
    Licensed to the Apache Software Foundation (ASF) under one or more
    contributor license agreements.  See the NOTICE file distributed with
@@ -15,7 +14,6 @@
    See the License for the specific language governing permissions and
    limitations under the License.
 ==================================================================== */
-        
 
 package org.apache.poi.hssf.usermodel.examples;
 
@@ -32,26 +30,23 @@ import java.util.Date;
  *
  * @author Glen Stampoultzis (glens at apache.org)
  */
-public class CreateDateCells
-{
-    public static void main(String[] args)
-        throws IOException
-    {
+public class CreateDateCells {
+    public static void main(String[] args) throws IOException {
         HSSFWorkbook wb = new HSSFWorkbook();
         HSSFSheet sheet = wb.createSheet("new sheet");
 
         // Create a row and put some cells in it. Rows are 0 based.
-        HSSFRow row = sheet.createRow((short)0);
+        HSSFRow row = sheet.createRow(0);
 
         // Create a cell and put a date value in it.  The first cell is not styled as a date.
-        HSSFCell cell = row.createCell((short)0);
+        HSSFCell cell = row.createCell(0);
         cell.setCellValue(new Date());
 
         // we style the second cell as a date (and time).  It is important to create a new cell style from the workbook
         // otherwise you can end up modifying the built in style and effecting not only this cell but other cells.
         HSSFCellStyle cellStyle = wb.createCellStyle();
         cellStyle.setDataFormat(HSSFDataFormat.getBuiltinFormat("m/d/yy h:mm"));
-        cell = row.createCell((short)1);
+        cell = row.createCell(1);
         cell.setCellValue(new Date());
         cell.setCellStyle(cellStyle);
 
@@ -59,6 +54,5 @@ public class CreateDateCells
         FileOutputStream fileOut = new FileOutputStream("workbook.xls");
         wb.write(fileOut);
         fileOut.close();
-
     }
 }
